Section completely 10 filled
Added the representation of a mindmap in plain text because asciidoc does not support the plantUML mindmap notation (at least i did not find how). Filled the quality scenarios making a a distinction between usage and change scenarios. Added a couple of attributes (as the template indicated) to extend the ones mentioned in 1.2